Fix bug related to large plugin instance files or large amount of files not being registered in the ChRIS DB

+++ chris_backend/plugininstances/models.py
@@ -38,7 +38,7 @@ class PluginInstance(models.Model):
     status = models.CharField(max_length=30, choices=STATUS_CHOICES, default='created')
     summary = models.CharField(max_length=4000, blank=True)
     raw = models.TextField(blank=True)
-    size = models.IntegerField(default=0)
+    size = models.BigIntegerField(default=0)
     error_code = models.CharField(max_length=7, blank=True)
     previous = models.ForeignKey("self", on_delete=models.CASCADE, null=True,
                                  related_name='next')
